Beautification
Added colors,days gain to currency and indices

+++ lib/models/stock.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class Stock{
   final String Name;
-  final int Price;
+  final double Price;
   final int Quantity;
 
   Stock({required this.Name,required this.Price,required this.Quantity});
